requirements/base.txt
Django==4.2.0
psycopg2-binary==2.9.9
python-environ==0.4.54
django-cors-headers==4.3.1
django-jazzmin==2.6.0
drf-yasg==1.21.7
djangorestframework==3.15.0
django-filter==24.1
django-ckeditor==6.5.1
django-phonenumber-field==7.0.2
phonenumbers==8.13.9
Pillow==9.1.1